set timeout and set home button

+++ src/app/device-only/services/home-button.service.ts
@@ -4,6 +4,7 @@ import { HomeService, AlertService, User, TokenService } from 'mychannel-shared-
 import { CreateOrderService } from './create-order.service';
 import { TransactionService } from 'src/app/shared/services/transaction.service';
 import { PriceOptionService } from 'src/app/shared/services/price-option.service';
+import { Router } from '@angular/router';
 @Injectable({
   providedIn: 'root'
 })
@@ -19,35 +20,42 @@ export class HomeButtonService {
     private transactionService: TransactionService,
     private priceOptionService: PriceOptionService,
     private tokenService: TokenService,
+    private router: Router
   ) {
     this.user = this.tokenService.getUser();
+    this.transaction = this.transactionService.load();
   }
 
   initEventButtonHome(): void {
     this.homeService.callback = () => {
-      this.alertService.question('ต้องการยกเลิกรายการขายหรือไม่ การยกเลิก ระบบจะคืนสินค้าเข้าสต๊อคสาขาทันที', 'ตกลง', 'ยกเลิก')
-        .then((response: any) => {
-          console.log(response);
-          if (response.value === true) {
-            this.createOrderService.cancelOrder(this.transaction).then((isSuccess: any) => {
-              console.log('isSuccess', isSuccess);
-              console.log('localtioncode', this.user.locationCode);
-              this.transactionService.remove();
-              this.priceOptionService.remove();
-              if (this.user.locationCode === '1213') {
-                window.location.href = '/smart-digital/main-menu';
-              } else {
-                window.location.href = '/';
-              }
-            });
-          }
-        }).catch((err: any) => {
-          console.log('err', err);
-          this.transactionService.remove();
-          this.priceOptionService.remove();
-          window.location.href = '/';
-        });
+      const url = this.router.url;
+      if (url.indexOf('result') !== -1) {
+        this.homeHandler();
+      } else {
+        this.alertService.question('ต้องการยกเลิกรายการขายหรือไม่ การยกเลิก ระบบจะคืนสินค้าเข้าสต๊อคสาขาทันที', 'ตกลง', 'ยกเลิก')
+          .then((response: any) => {
+            if (response.value === true) {
+              this.createOrderService.cancelOrder(this.transaction).then((isSuccess: any) => {
+                this.transactionService.remove();
+                this.priceOptionService.remove();
+                this.homeHandler();
+              });
+            }
+          }).catch((err: any) => {
+            this.transactionService.remove();
+            this.priceOptionService.remove();
+            this.homeHandler();
+          });
+      }
     };
+  }
+
+  homeHandler(): any {
+    if (this.user.locationCode === '1213') {
+      window.location.href = '/smart-digital/main-menu';
+    } else {
+      window.location.href = '/';
+    }
   }
 
 }
